+++ CHANGELOG.md
@@ -4,12 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [8.5.4] - 2018-07-31
 ### Added
 - Adds *Checkout Commit (via Terminal)* command (`gitlens.terminalCheckoutCommit`) to commit node(s) of the *GitLens* explorer &mdash; closes [#463](https://github.com/eamodio/vscode-gitlens/issues/463)
 
 ### Changed
-- Improves performance of blame annotation rendering
+- Improves performance of providing blame annotations, especially for large files (a ~33% improvement on some files)
 - Changes commit search by sha to jump directly to the commit (if one is found), rather than having to click through to the commit
 - Updates settings descriptions to better support the new settings editor in vscode
 - Changes *Rebase to Commit (via Terminal)* command (`gitlens.terminalRebaseCommit`) to apply to all commit node(s) of the *GitLens* explorer